scripts/spread_jcb/requirements.txt
kafka-python==2.0.2
pyyaml==5.4.1